Generate ssl certs on supervisor startup
