+++ bin/command_executor.py
@@ -1,0 +1,25 @@
+#! /usr/bin/python
+import sys
+import subprocess
+
+def write_stdout(s):
+    sys.stdout.write(s)
+    sys.stdout.flush()
+
+def write_stderr(s):
+    sys.stderr.write(s)
+    sys.stderr.flush()
+
+def main(args):
+    write_stdout('READY\n') # transition from ACKNOWLEDGED to READY
+    line = sys.stdin.readline()  # read header line from stdin
+    write_stderr(line) # print it out to stderr
+    headers = dict([ x.split(':') for x in line.split() ])
+    data = sys.stdin.read(int(headers['len'])) # read the event payload
+    res = subprocess.call(args, stdout=sys.stderr); # don't mess with real stdout
+    write_stderr(data)
+    write_stdout('RESULT 2\nOK') # transition from READY to ACKNOWLEDGED
+
+if __name__ == '__main__':
+    main(sys.argv[1:])
+    import sys
