+++ etc/supervisor/conf.d/envoy.ini
@@ -1,26 +1,62 @@
+# Supervisor does not support starting a program after another has finished.
+# To work around this, we define all the programs that should run, and
+# listeners to listen for one to exit, then call the next one.
+# Only the first program in the chain should have autostart=true.
+
+##################
+# Envoy Programs #
+##################
+
+# Run the configuration script, this will interpolate and write the config files for Envoy
 [program:envoy_config]
 command=/usr/local/bin/configure.py /etc/envoy/
 startsecs=0
-priority=900
+autorestart=false
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
 stderr_logfile=/dev/stderr
 stderr_logfile_maxbytes=0
 
+# Generate some self signed certs for Envoy to use with its SSL configuration
+[program:envoy_gen_certs]
+command=/usr/local/bin/generate_certs.sh
+startsecs=0
+autostart=false
+autorestart=false
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+stderr_logfile=/dev/stderr
+stderr_logfile_maxbytes=0
+
+# Envoy itself
 [program:envoy]
-command=/usr/local/bin/envoy -c /etc/envoy/envoy.yaml --v2-config-only --service-cluster c0 --service-node n0 --log-level warn
+command=/usr/local/bin/envoy -c /etc/envoy/envoy.yaml --v2-config-only --service-cluster c0 --service-node n0 --log-level %(ENV_ENVOY_LOG_LEVEL)s
 autostart=false
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
 stderr_logfile=/dev/stderr
 stderr_logfile_maxbytes=0
 
-[eventlistener:envoy_exit]
-command=/usr/local/bin/kill_supervisor.py
-process_name=envoy
-events=PROCESS_STATE_FATAL
+###################
+# Envoy Listeners #
+###################
 
+# Generate certs once the config files have been written
 [eventlistener:envoy_configured]
-command=supervisorctl start envoy
+command=command_executor.py supervisorctl start envoy_gen_certs
+autorestart=false
 process_name=envoy_config
 events=PROCESS_STATE_EXITED
+
+# Start Envoy after the certs have been generated
+[eventlistener:envoy_certs_generated]
+command=command_executor.py supervisorctl start envoy
+autorestart=false
+process_name=envoy_gen_certs
+events=PROCESS_STATE_EXITED
+
+# Stop supervisor if Envoy exits or can't run (this will also stop the container)
+[eventlistener:envoy_exit]
+command=command_executor.py supervisorctl shutdown
+process_name=envoy
+events=PROCESS_STATE_FATAL
