+++ etc/supervisor/supervisord.conf
@@ -25,7 +25,7 @@ file=/tmp/supervisor.sock   ; the path to the socket file
 ;password=123               ; default is no password (open server)
 
 [supervisord]
-logfile=/dev/stdout              ; main log file; default $CWD/supervisord.log
+logfile=/dev/null                ; main log file; default $CWD/supervisord.log
 logfile_maxbytes=0               ; max main logfile bytes b4 rotation; default 50MB
 logfile_backups=0                ; # of main logfile backups; 0 means none, default 10
 loglevel=info                    ; log level; default info; others: debug,warn,trace
